Form za unos line chart datasetova

+++ scripts/prepare.sh
@@ -37,12 +37,29 @@ fi
 
 if [ ! "$CI" ]; then
   printf "[Mess] Setting up secrets with 'dotnet'...\n"
+
+  if [ ! -f "$ROOT_DIR/secrets.json" ]; then
+    printf "[Mess] File 'secrets.json' not found.\n"
+    printf "[Mess] Please copy over secrets to the worktree.\n"
+    printf "\n"
+    exit 1
+  fi
+
+  if [ ! -f "$ROOT_DIR/secrets.sh" ]; then
+    printf "[Mess] File 'secrets.sh' not found.\n"
+    printf "[Mess] Please copy over secrets to the worktree.\n"
+    printf "\n"
+    exit 1
+  fi
+
   dotnet user-secrets set --project "$ROOT_DIR/src/Mess.Web/Mess.Web.csproj" <"secrets.json"
+
   if [ ! -d "$ROOT_DIR/secrets" ]; then
     mkdir "$ROOT_DIR/secrets"
   fi
   cp "$ROOT_DIR/secrets.json" "$ROOT_DIR/secrets"
   cp "$ROOT_DIR/secrets.sh" "$ROOT_DIR/secrets"
+
   printf "\n"
 fi
